Extract common code into mixins
